Updated sum in counter to be a double

+++ src/application/demo.h
@@ -55,7 +55,7 @@ public:
   void counter() {
     pln("Counter Started");
     DataFrame* v = kv_->waitAndGet(main);
-    size_t sum = 0;
+    double sum = 0;
     for (size_t i = 0; i < 100*1000; ++i) sum += v->get_double(0,i);
     p("The sum is  ").pln(sum);
     DataFrame::fromScalar(verify, kv_, sum);
